+++ Content.Benchmarks/DeviceNetworkingBenchmark.cs
@@ -6,6 +6,7 @@ using Content.IntegrationTests.Pair;
 using Content.IntegrationTests.Tests.DeviceNetwork;
 using Content.Server.DeviceNetwork;
 using Content.Server.DeviceNetwork.Systems;
+using Content.Shared.DeviceNetwork;
 using Robust.Shared;
 using Robust.Shared.Analyzers;
 using Robust.Shared.GameObjects;
