+++ Content.Client/DeviceLinking/DeviceLinkSystem.cs
@@ -1,0 +1,8 @@
+using Content.Shared.DeviceLinking;
+
+namespace Content.Client.DeviceLinking;
+
+public sealed class DeviceLinkSystem : SharedDeviceLinkSystem
+{
+
+}
