+++ Content.IntegrationTests/Tests/DeviceNetwork/DeviceNetworkTest.cs
@@ -2,6 +2,7 @@ using System.Numerics;
 using Content.Server.DeviceNetwork;
 using Content.Server.DeviceNetwork.Components;
 using Content.Server.DeviceNetwork.Systems;
+using Content.Shared.DeviceNetwork;
 using Robust.Shared.GameObjects;
 using Robust.Shared.IoC;
 using Robust.Shared.Map;
