+++ Content.Server/Atmos/Piping/Unary/EntitySystems/GasVentPumpSystem.cs
@@ -16,6 +16,7 @@ using Content.Shared.Atmos.Monitor;
 using Content.Shared.Atmos.Piping.Unary.Components;
 using Content.Shared.Atmos.Visuals;
 using Content.Shared.Audio;
+using Content.Shared.DeviceNetwork;
 using Content.Shared.Examine;
 using Content.Shared.Tools.Systems;
 using JetBrains.Annotations;
