+++ Content.Server/DeviceLinking/Events/SignalReceivedEvent.cs
@@ -1,4 +1,5 @@
 using Content.Server.DeviceNetwork;
+using Content.Shared.DeviceNetwork;
 
 namespace Content.Server.DeviceLinking.Events;
