+++ Content.Server/SensorMonitoring/BatterySensorSystem.cs
@@ -1,6 +1,7 @@
 ﻿using Content.Server.DeviceNetwork;
 using Content.Server.DeviceNetwork.Systems;
 using Content.Server.Power.Components;
+using Content.Shared.DeviceNetwork;
 
 namespace Content.Server.SensorMonitoring;
