+++ Content.Server/SensorMonitoring/SensorMonitoringConsoleSystem.cs
@@ -10,6 +10,7 @@ using Content.Server.Power.Generation.Teg;
 using Content.Shared.Atmos.Monitor;
 using Content.Shared.Atmos.Piping.Binary.Components;
 using Content.Shared.Atmos.Piping.Unary.Components;
+using Content.Shared.DeviceNetwork;
 using Content.Shared.DeviceNetwork.Components;
 using Content.Shared.DeviceNetwork.Systems;
 using Content.Shared.SensorMonitoring;
